# server/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Service API
  version: 1.1.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vpn:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
    VpnCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    VpnUpdateRequest:
      type: object
      properties:
        name:
          type: string
...
  responses:
    TooManyRequests:
      description: Too many requests
      headers:
        Retry-After:
          description: When to retry the request
          schema:
            type: string
...
paths:
...
  /api/vpn:
    get:
      summary: List VPNs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: VPN list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vpn'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      summary: Create VPN
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VpnUpdateRequest'
      responses:
        '201':
          description: Created VPN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vpn'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/vpn/{id}:
    patch:
      summary: Update VPN
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VpnCreateRequest'
      responses:
        '200':
          description: Updated VPN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vpn'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      summary: Delete VPN
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted VPN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vpn'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...